fix: change route after click detail
