+++ src/App.css
@@ -7,7 +7,6 @@
   width: 100%;
   max-width: 1280px;
   margin: 0 auto;
-  padding: 0 15px;
   text-align: center;
 }
