+++ src/pages/DetailPage.tsx
@@ -3,7 +3,6 @@ import { useLoaderData, useNavigation, Link } from "react-router-dom";
 import { useState, useEffect } from "react";
 import { Character } from "../types/types";
 import Loader from "./../components/Loader";
-import { LoaderFunctionArgs } from "react-router-dom";
 
 const DetailPage = () => {
   // считывает  loader: detailLoader из routes
@@ -72,19 +71,3 @@ const DetailPage = () => {
 };
 
 export default DetailPage;
-
-
-
-export const detailLoader = async ({ params }: LoaderFunctionArgs) => {
-  await new Promise((resolve) => setTimeout(resolve, 200));
-  const { id } = params;
-  const response = await fetch(`https://rickandmortyapi.com/api/character/${id}`);
-  if (response.status === 404) {
-    throw new Error("Not Found");
-  }
-  if (!response.ok) {
-    throw new Error("Network response was not ok");
-  }
-  const data = await response.json();
-  return data;
-};
